Odds Api Swagger - Controllers code cleaning + OddsUtils class
